Addresses PR comments Signed-off-by: Sima Nadler <sima@il.ibm.com>
